Fixed bad bug!

+++ modules/miscutil/lib/intbitset.h
@@ -23,30 +23,30 @@
 typedef unsigned long long int word_t;
 typedef unsigned char bool_t;
 
-extern const size_t wordbytesize;
-extern const size_t wordbitsize;
+extern const int wordbytesize;
+extern const int wordbitsize;
 
 typedef struct {
     int size;
-    size_t allocated;
+    int allocated;
     word_t universe;
     int tot;
     word_t *bitset;
 } IntBitSet;
 
-IntBitSet *intBitSetCreate(register const size_t size, const bool_t universe);
-IntBitSet *intBitSetCreateFromBuffer(const void * const buf, const size_t bufsize);
-IntBitSet *intBitSetResetFromBuffer(IntBitSet *const bitset, const void *const buf, const size_t bufsize);
+IntBitSet *intBitSetCreate(register const int size, const bool_t universe);
+IntBitSet *intBitSetCreateFromBuffer(const void * const buf, const int bufsize);
+IntBitSet *intBitSetResetFromBuffer(IntBitSet *const bitset, const void *const buf, const int bufsize);
 IntBitSet *intBitSetReset(IntBitSet *const bitset);
 void intBitSetDestroy(IntBitSet *const bitset);
 IntBitSet *intBitSetClone(const IntBitSet * const bitset);
-size_t intBitSetGetSize(IntBitSet * const bitset);
-size_t intBitSetGetAllocated(const IntBitSet * const bitset);
-size_t intBitSetGetTot(IntBitSet * const bitset);
-void intBitSetResize(IntBitSet *const bitset, register const size_t allocated);
-bool_t intBitSetIsInElem(const IntBitSet * const bitset, register const size_t elem);
-void intBitSetAddElem(IntBitSet *const bitset, register const size_t elem);
-void intBitSetDelElem(IntBitSet *const bitset, register const size_t elem);
+int intBitSetGetSize(IntBitSet * const bitset);
+int intBitSetGetAllocated(const IntBitSet * const bitset);
+int intBitSetGetTot(IntBitSet * const bitset);
+void intBitSetResize(IntBitSet *const bitset, register const int allocated);
+bool_t intBitSetIsInElem(const IntBitSet * const bitset, register const int elem);
+void intBitSetAddElem(IntBitSet *const bitset, register const int elem);
+void intBitSetDelElem(IntBitSet *const bitset, register const int elem);
 bool_t intBitSetEmpty(const IntBitSet * const bitset);
 IntBitSet *intBitSetUnion(IntBitSet *const x, IntBitSet *const y);
 IntBitSet *intBitSetXor(IntBitSet *const x, IntBitSet *const y);
@@ -56,7 +56,7 @@ IntBitSet *intBitSetIUnion(IntBitSet *const dst, IntBitSet *const src);
 IntBitSet *intBitSetIXor(IntBitSet *const dst, IntBitSet *const src);
 IntBitSet *intBitSetIIntersection(IntBitSet *const dst, IntBitSet *const src);
 IntBitSet *intBitSetISub(IntBitSet *const x, IntBitSet *const y);
-size_t intBitSetAdapt(IntBitSet *const x, IntBitSet *const y);
+int intBitSetAdapt(IntBitSet *const x, IntBitSet *const y);
 int intBitSetGetNext(const IntBitSet *const x, register int last);
 /** Compare.
  * Compare two intbitset.
